backend/requirements.txt
# Agentic Inbox & Meeting Follow-up App — Backend
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
alembic>=1.13.0,<2.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.0,<2.0.0
httpx>=0.26.0,<1.0.0
redis>=5.0.0,<6.0.0
celery>=5.3.0,<6.0.0
google-auth>=2.27.0,<3.0.0
google-auth-oauthlib>=1.2.0,<2.0.0
google-api-python-client>=2.114.0,<3.0.0
msal>=1.26.0,<2.0.0
beautifulsoup4>=4.12.0,<5.0.0
bleach>=6.1.0,<7.0.0
python-multipart>=0.0.6
tenacity>=8.2.0,<9.0.0
structlog>=24.1.0,<25.0.0
openai>=1.10.0,<2.0.0
anthropic>=0.18.0,<1.0.0
tiktoken>=0.6.0,<1.0.0
boto3>=1.34.0,<2.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx>=0.26.0
